Add documentation tests
