+++ Cargo.toml
@@ -32,6 +32,7 @@ bs58 = "0.4.0"
 byteorder = "1.5.0"
 cadence-macros = "1.2.0"
 clap = {"version" = "4.5.2", features = ["derive"]}
+dirs = "5.0.1"
 env_logger = "0.10.0"
 futures = "0.3.30"
 hyper = "0.14.23"
