+++ src/lib.rs
@@ -1,6 +1,7 @@
 // Required for capturing backtraces
 pub mod api;
+pub mod common;
 pub mod dao;
 pub mod ingester;
 pub mod migration;
-pub mod common;
+pub mod openapi;
