+++ tests/integration_tests/open_api_tests.rs
@@ -1,0 +1,6 @@
+use photon_indexer::openapi::update_docs;
+
+#[test]
+pub fn test_documentation_generation() {
+    update_docs(true);
+}
